pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yenpoint_1satordinals"
version = "0.1.0"
description = "A library for creating 1Sat Ordinals NFTs on BSV blockchain"
authors = [
    {name = "Ken Sato", email = "info@yenpoint.jp"},
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "bsv-sdk>=0.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["yenpoint_1satordinals*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"